+++ geodecoding/data/disabled_sports_classes_not_found.csv
@@ -1,0 +1,7 @@
+﻿BSNS_NO,FCLTY_NM,ITEM_CD,ITEM_NM,CTPRVN_CD,CTPRVN_NM,SIGNGU_CD,SIGNGU_NM,FCLTY_ADDR,FCLTY_DETAIL_ADDR,ZIP_NO,TEL_NO,COURSE_NM,TROBL_TY_NM,COURSE_NO,COURSE_ESTBL_YEAR,COURSE_ESTBL_MT,COURSE_BEGIN_DE,COURSE_END_DE,COURSE_REQST_NMPR_CO,COURSE_PRC
+1258205765,안성시국민체육센터,12,수영,41,경기,41550,안성시,경기도 안성시 보개면 종합운동장로 162,안성시시설관리공단,17508,0316510741,수영(자유수영-월이용) -성인,,209034,2024,09,20240901,20240930,1,20000.00000
+1258205765,안성시국민체육센터,25,헬스,41,경기,41550,안성시,경기도 안성시 보개면 종합운동장로 162,안성시시설관리공단,17508,0316510741,헬스(자유헬스 -월이용)- 성인,,209040,2024,09,20240901,20240930,1,17500.00000
+2159517545,꾸러기태권도,22,태권도,30,대전,30140,중구,대전광역시 중구 수도산로 28,3층 꾸러기태권도(대흥동),34937,0422223342,꾸러기태권도,,4708014,2024,09,20240901,20240930,1,110000.00000
+2159517545,꾸러기태권도,22,태권도,30,대전,30140,중구,대전광역시 중구 수도산로 28,3층 꾸러기태권도(대흥동),34937,0422223342,꾸러기태권도,,4708014,2024,10,20241001,20241031,1,110000.00000
+5039180933,호키스피드체육관,22,태권도,27,대구,27290,달서구,대구광역시 달서구 와룡로53길 7~105,3층,42631,025230555,태권도,,4624014,2024,09,20240901,20240930,2,110000.00000
+5039180933,호키스피드체육관,22,태권도,27,대구,27290,달서구,대구광역시 달서구 와룡로53길 7~105,3층,42631,025230555,태권도,,4624014,2024,10,20241001,20241031,2,110000.00000
